AI/cloud/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
openai>=1.30.0
python-multipart>=0.0.20
requests>=2.32.3
python-dotenv>=1.0.1
transformers==4.45.1
peft>=0.7.0
torch>=2.1.0
accelerate>=0.25.0
faster-whisper>=1.0.0